producers/requirements.txt
kafka-python-ng==2.2.2
faker==18.9.0